leds: leds-mlxreg: Downstream: Send udev event from leds-mlxreg driver.
Add sending udev event from leds-mlxreg driver in case of color change.
This patch will not be accepted in upstream as it provides no regular flow.
It's done as a workaround of SN2201 limitation, difference, and in order
to save compatibility of NOS work with sysfs attributes.
The patch should be taken to NOSs as exceptions to the upstream procedure.

Signed-off-by: Michael Shych <michaelsh@nvidia.com>