hwmon: (mlxsw): Downstream: Allow fan speed setting  granularity of 1 PWM
Currently PWM setting is allowed with 10 percent stepping.
Such configuration is aligned with thermal drivers, which are used to be
bound to "mlxsw" driver.

This binding happens when the cooling instances created by the driver are
bound to some kernel thermal driver.

In case system is not using kernel thermal control and the cooling
instances created by the driver are not bound to any thermal driver, the
driver still does not allow setting of PWM granularity less than 10
percent.

Allow setting fan with one percent granularity, thus any user space
thermal application will be able to set PWM to any allowed value in range
from 51 PWM to 255 PWM.

Note: this is downstream patch, since it can affect functionality for
the Nvidia users running kernel thermal control. So, it is not going to be
submitted to up-stream.

Signed-off-by: Oleksandr Shamray <oleksandrs@nvidia.com>
Reviewed-by: Vadim Pasternak <vadimp@nvidia.com>