mlxsw: core_hwmon: Fix module sensor number for QM3200
This patch hardcodes the number of modules per ASICs on QM3200 systems
instead of reading the actual number of modules from the MGPIR register.
This is a WA for recent FW versions that incorrectly report the number
of modules per ASIC.

On QM3200 systems ASIC1 and ASIC2 control 18 and 19 modules respectively.
ASICs are identified by thir i2c bus number, with ASIC1 using i2c bus 2.

Signed-off-by: Felix Radensky <fradensky@nvidia.com>