platform/mellanox: mlxreg-hotplug: Add support for handling interrupt storm
In case of broken hardware, it is possible that broken device will
flood interrupt handler with false events.
For example, if fan or power supply has damaged presence pin, it will
cause permanent generation of presence in / presence out events.
As a result, interrupt handler will consume a lot of CPU resources and
will keep raising "UDEV" events to the user space.

At the same device with damaged pin still will be capable to provide
telemetry date.

Provide mechanism allowing to detect device causing interrupt flooding
and mask interrupt for this specific device, to isolate from interrupt
handling flow.

Use the following criteria: if the specific interrupt was generated 'N'
time during 'T' seconds, such device is to be considered as broken and
will be closed for getting interrupts.

User will be notified through the log error and will be instructed to
replace broken device.

Signed-off-by: Roy Cohen <roycohen@nvidia.com>